// src/core/api/generated/spring/schemas/SystemConfigDTOStatus.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type SystemConfigDTOStatus = typeof SystemConfigDTOStatus[keyof typeof SystemConfigDTOStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SystemConfigDTOStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  ARCHIVED: 'ARCHIVED',
} as const;